Adding better support for filtering change set entries in memory
When querying for drift entries by criteria and when the criteria
specifies filters that include at least one of the drift entry fields,
we have to filter the result set again once it is returned from the
database. This is due to the schema design I have chosen. In the MongoDB
design for drift, a change set along with its entries are stored as a
single document in the db. If you query for change set drift entries,
the datbase (by default) returns the entire document which may include
drift entries that should be exlucded from the result; thus, we need to
filter again after the results are returned from the db.