Adding a new change set header, "repeated"
The repeated flag indicates whether or not the change set is a repeat of
a previous change set. This is in support of pinned snapshots where the
same drift can be detected repeatedly. When we detect the same drift
back to back during two drift detection runs, we do not want to
increment the current snapshot version the second time, but we do
however want to report to the server that there is still drift. This new
header faciliates that.