Updating logic for drift detection with a pinned snapshot
When using a pinned snapshot and when drift is detected, the agent now
checks to see if the set of deltas is the same deltas from the previous
change set. This can happen with a pinned snapshot because the snapshot
used for comparison does not get updated; consequently, the same drift
can get detected over and over again. If the agent determines that the
deltas are the same, it notifies the server that the same change set has
been detected. No change set file however is sent to the server. The
server in turn will process any drift alert definitions but no new
change set is persisted.