Updating inventory sync logic to handle pinned snapshots
This commit includes logic for server sync at start up. Additional logic
is needed to handle runtime syncs (i.e., syncs after the PC is fully
initialized). For a pinned drift definition, the current snapshot, the
pinned snapshot, and a snapshot for the latest change set have to be
fetched from the server and written out to disk.