Initial commit for DriftDefinitionTemplate
Up until now drift templates were represented and stored as
ConfigurationTemplates.. We have decided to turn it into an entity to
better support the functionality that needs to be implementing around
templates.

I have also changed AbstractEJB3Test.shutdownEmbeddedJboss to run as an
@AfterSuite method instead of as an @AfterGroup method. A while back I
changed it to an @AfterGroup method to support running Unitils-based
tests in the same suite. Unitils is not being used and the @AfterGroup
was causing problems for me in DriftDefinitionTemplateTest. In
DriftDefinitionTemplateTest I added an @AfterClass method to clean up
the database, but it was running after the embedded container shut down.
Changing shutdownEmbeddedJboss to an @AfterSuite method resolved the
issue.