Adding assocation between DriftDefinition and DriftDefinitionTemplate
This commit updates the dbsetup and dbupgrade scripts by adding a FK in
the rhq_drift_config to rhq_drift_def_template.

A unidirectional one-to-many asscociation has been added in
DriftDefinition. Not sure if we need to make the association
bi-directional.

Updated the wizard in coregui to set the template on the newly created
definition.