BZ 743632, BZ 634648
634648 - Removed the JSF-based metric time range picker and replaced it with
the GWT based one even on the JSF pages (compositing the page from a JSF
iframe and the GWT picker).

743632 - Made the group tables view refreshable. Only the metrics table is
being refreshed so that the list of members isn't reloaded when merely
updating the metric time range. Added the "Refresh" buttons to both
the metric table and the member table for the page to be inline with the
resource-specific counterpart that does show them, too.