Updating javadocs and adding logic for accessing drifts of the initial change set
If the change set is the initial change set, i.e., version zero, then
its drifts are to be accessed through the field initialDriftSet. As part
of the design changes to support a shared pinned snapshot among multiple
drift definitions, the initial change set will now always access its
drifts through a JPADriftSet.

Going through JPADriftSet is necessary in the case of a pinned template.
It is worth noting that it is not necessary to go through JPADriftSet in
the case of an unpinned template; however, we are doing so to keep
things consistent.

Lastly, it should be noted all of the details around JPADriftSet are
implementation details of the rhq drift server plugin and are not
applicate to other plugin implementations.