Updating logic for querying change sets.
When we are fetching the initial change set, which is a full snapshot,
we load the drifts through JPADriftSet object. Drifts for delta change
sets are still loaded in the same way that they have previously been
loaded.

The semantics of findDriftChangeSetsByCriteria is changed slightly as
well. Unless specified, implementations can assume that the query is
only for delta change sets. And if the caller wants to retrive both the
initial change set, and delta change sets (as is the case when building
a snapshot), then two separate calls need to be made to
findDriftchangeSetsByCriteria.