Attemmpting to fix test failure
This failure is only happening on hudson. I have seen this before. A
NoClassDefFoundError is getting throw from tests in server/jar that call
JPAUtil.executeInTransaction. I am not sure why this is happening. It
does not happen locally, and I use the executeInTransaction elsewhere in
core/domain.

I am adding a executeInTransaction method to AbstractEJBTest that is a
wrapper around the call to the one in JPAUtil. This is what is done in
the AbstractEJB3Test in core/domain. Hoping it will get past this error.