Attempting to fix test failure that may be due to a surefire bug
I maanged to reproduce the failure that is happening on jenkins but only
when running all tests in server/jar. This provided some additional
insight that allowed me to eliminate jenkins as the source of the
problem. I think the problem may be related to
http://jira.codehaus.org/browse/SUREFIRE-313 so I am upgrading the
version of the surefire plugin to 2.4 which is the version that contains
the fix for SUREFIRE-313.