Adding server side logic for pinning snapshots
Pinning a snapshot will result in the following on the server:

  1) The snapshot is made the initial change set for the drift def
  2) All change sets for the drift def are deleted
  3) The pinned flag on the drift def is set
  4) A pin snapshot request is sent to the agent

This commit adds a new method to DriftServerPluginFacet for persisting
the snapshot. This commit also introduces a lot of new test
infrastructure. DriftServerTest is a base test class that encapsulates
common set up/tear down logic for the database, and it also starts up
the server plugin container(s) so that tests can call through to the
plugin. This will facilitate testing other drift plugin implementations
as well.