Call EntityManager.persist to get id for template returned from createTemplate
Adding call to EntityManager.persist(DriftDefinitionTemplate) so that
the id gets set on the template before it is returned.

This is also the initial commit for ManageDriftDefinitionsTest. Right
now the primary focus of this test class is pinning definitions and
creating definitions from pinned templates. There is some overlap
between this class and DriftManagerBeanTest. I started out refactoring
DriftManagerBeanTest to extend DriftServerTest, but that resulted in
more of the ClassNotFoundExceptions on inner classes. This is a
surefire/TestNG problem that I don't want to spend any more time on
right now. I do not get the exceptions though with the new tests in
ManageDriftDefinitionsTest.