Addding support for generating snapshot for definition created from pinned template
When a drift definition is created from a pinned template, the
definition will also be pinned and its initial change set is copied from
the template. This commit generates a snapshot of that initial change
set and sends it to the agent along with the definition so that drift
detection can get started. The agent writes the snapshot to disk.

There are a couple things that still need to get worked out, but I am
committing this now to resolve a test failure in DriftManagerBeanTest.