Adding support for generating snapshot that is not tied to a defnintion
Up until now, the code for generating snapshots assumes that the change
sets that comprise the snasphot are tied to a particular definition. But
with a template snapshot, there is no definition to which the change
sets belong.

With this commit when you create a definition from a pinned template the
server now sends a populated snapshot to the agent. You have to manually
set the pinned flag on the definition though. That'll get fixed in a
subsequent commit.