fixing logic for generating snapshot when creating drift definition
In my previous commit I created a snapshot from the template's pinned
change set which is counter to what we have been doing with snapshots to
date. I figured out why I was generating empty snapshots off of the
drift definition. I was going across transactionial boundaries when
calling getSnapshot. I changed it so that no new transaction would be
created and reverted the snapshot logic back to always being specific to
a particular drift definition.