Only look at directory specs when generating list of templates for wizard
When the pin template wizard launches, we should at the base directory
specifications which includes the base directory itself and filters to
determine the eligible set of templates to which a snapshot can be
pinned. The other fields in a definition are editable so it is natural
that they may vary from the template.