Set the template reference of definitions when pinning a template
Need to make sure to set the template reference for each DriftDefinition
that is updated/replaced when a template is pinned.

Adding a new test to verify that attached definitions get pinned when a
template gets pinned.

Also adding support for sharing database fixtures between multiple tests
with the @InitDB annotation. The pinTemplate test was getting really big
so I decided to break out the verification steps into separate tests.
These tests have @InitDB(false) so that they can reuse the template and
definitions created in the pinTemplate method.