Fixing bug in resource-level pinning
There was a bug with pinning a snapshot to a definition where the
definition was always getting pinned to the current snapshot. The
correct behavior is that it should get pinned to whichever snapshot the
user selects. This was a minor fix in that DriftManager.pinSnapshot was
ignoring the snapshot argument.

The rest of this commit involves adding back the resourceId param to the
DriftAgentService methods. I am doing this because the DriftDefinition
getting sent to the agent might not have its resource field/proxy
initialized.