fixing failing tests
Tests in AvailabilityManagerTest were failing because of a call in
AgentManagerBean to load the system settings. A list of the drift server
plugins is stored as part of the system settings. To obtain that list
SystemManagerBean attempts to access DriftServerPluginContainer via the
server plugin service.

I have supplied AvailabilityManagerTest with a stubbed out server plugin
service so that the SystemManagerBean is able to load the default
settings without error.