Initial commit for integration-tests module
The server/integration-tests module was born out of the need to resolve
a dependency issue for drift tests. There were drift tests in server/jar
that depended on having a drift server plugin installed, but the server
plugins are not built until after server/jar is built.

This commit moves tests over to the integrations-tests module which is
built after server/jar and the plugins, thus resolving the dependency
issue.

server/jar now produces a test jar artifact so that integration-tests
has access to AbstractEJB3Test. A better, longer term solution would be
to extract AbstractEJB3Test out into a common module since it is
duplicated in core/domain as well. I don't have time for that right now
though.