BZ 687992 - A final touch on making the apache config file parsing correct. The parser now keeps track of the "current" ServerRoot (which is defaulted to the value hardcoded in the binary at the start and then updated each time a ServerRoot directive is encountered during the config file parsing).
Also made some minor refactoring to consolidate and reuse the runtime
configuration extracting routines.

A final update is to use the runtime configuration to figure out
the paths of the httpd executable as well as control script path (before
we used the full config and so could potentially get wrong results if
ServerRoot was hidden in some conditional blocks).

The runtime config is now also used when figuring out the sample address of
a virtual host during its creation.