[BZ 749126] - Loosening validation requirements on import. This commit is quite big due to the refactoring but does only a couple of simple things:
1) The SynchronizationManagerRemote.importAllSubsystems now returns
an ImportReport which contains the notes from subsystem importers about
"interesting" events during the import. This is used to report warnings
or other info to the caller of the import.
Namely the MetricTemplateImporter now reports all the metric templates
from the export file that were not present in the target database
(which basically means that source and target RHQ install had different
plugins installed).

2) Import is lenient about unknown validator classes (this is to not choke
on "old" import files that still may be perfectly valid).

3) The former ConsistencyValidator has been split into 2 separate concepts:
   ConsistencyValidator - persists its config into the export file, checks
   the consistency of the target RHQ install as a whole.
   EntityValidator - configured by the importer, does NOT persist anything
   into the export file, used to validate individual entities in the export
   file before the import begins.

4) Some new unit tests and a couple of simple changes in the code to
   support better testability.

5) DeployedAgentPluginsValidator was "decommissioned" but left in
   the codebase as an example implementation.