[BZ 753000] Normalize on using forward slashes in path names
To avoid issues with back slashes we are normalizing on using forward
slashes in path names even for windows. FilterFileVisitor uses the
method FilenameUtils.normalize from apache commons. This method is used
to normalize relative paths. We discovered that it also normalizes the
path separator such that it will convert path separators to be the
system default. That call is now wrapped to replace back slashes with
forward slashes.