[BZ 749419] Do not allow pinned definition to be re-pinned
When a definition is created from a pinned template, with this commit we
do not allow the definition to be repinned. The "Pin to Definition" and
"Pin to Template" buttons in the snapshot view are disabled. And calling
DriftManagerBean.pinSnapshot from the remote API now performs a check
and will throw an IllegalArgumentException.