[BZ 749419] We still want to allow pinning a snapshot to a template
Pinning a snapshop to a template only effects the definition when the
template is the one from which the definition was created. If you pin to
a new template or some other existing template, the definition (to which
the snapshot belongs) remains unmodified.

We do want to allow "repinning" particularly to support planned changes.
Suppose you have a pinned template and pinned definition for an EAP
server. As part of a planned change, you deploy a new or updated
application to the EAP server. Drift monitoring is set up for the EAP
server. A new snapshot is generated after the application is deployed,
and you will want to repin the template to that new snapshot.