[BZ 753827] Do not allow templates to be created with duplicate names
Template names for a given resource type must be unique. This commit has
three changes.

1) Add validation check in DriftTemplateManagerBean.createTemplate. An
IllegalArgumentException is thrown if the name is a duplicate. This
handles calls through the remote api as well as through the UI.

2) Add a validation check to the wizard in the drift templates view. If
you specify a duplicate name, the name field is flagged as being in
error when you click finish. The hover text says that the name must be
unique.

3) Add the same validation check as in 2) to the wizard for pinning a
snapshot to a template.