[BZ 753827] Fixing regression for pinning snapshot to existing template
My fix for bug 753827 caused a regression that breaks the wizard for
pinning a snapshot to an existing template. I was previously unware that
if you selected a new template in the wizard, that the code still runs
through the config step. In that case, the config editor in the config
step page is never intialized and this caused an NPE as a result of my
changes. A check is now in place to determine whether or not the
validation needs to be performed, and thus avoiding the NPE.