Adding method to retrieve drift content as byte array
Yesterday I discovered an issue while working on a CLI script that was
trying to fetch drift file content. Whitespace characters like \n and \t
were getting stripped out and replaced with a single whitespace
character. After a lot of debugging I determined that this is happening
in the transport layer which is jboss remoting. Fetching file content as
a byte array and then constructing a string on the client side bypasses
this issue.