Handle agent being down when pinning snapshot to definition
Prior to this commit when you pinned a snapshot to a definition, an
error would be reported in the UI if the agent was down or otherwise
unreachable; however, the server side operation completed successfully.
This commit adds exception handling and logging to deal with the agent
being unreachable.

This commit includes another critical change which is updating the
resource's mtime. Without updating the mtime, the definition's update is
not picked up during inventory sync when the agent restarts.