[BZ 756100] Use file timestamp and size during drift detection
This commit introduces an enhancement to drift detection by reducing the
amount of digest calculation that is performed. Digest calculation is
CPU-intensive.

The change set parser code has been updated to read/write file timestamp
and size. DriftDetector looks at the addition file info now to determine
if there has been a change. If the timestamp and size have not changed,
DriftDetector will skip recalculating the SHA. This results in a
significant speed up in drift detection.

Since the additional fields are not persisted on the server, there may
be times when the file timestamp and size are not available. This can
happen for example when the agent is restarted and the data directory is
purged. The agent has to download snapshots from the server at that
point, and they will not have the file modification time or size.