[Bug 760289 - Excessive file scanning in drift detection when using includes filters] Now, when using includes file paths limit the directory scanning to only those included directories.
Note that using a "." as an includes path basically translates to using
the base directory, in which case the scan will be as it is now.
A future enhancement may be to analyze the pattern and decide
whether a recursive scan is necessary.  Currently. So, using includes
patterns to just look for certain files in the base directory will
expose you to the full scan.