DBInstallationTest needs to use its own dedicated database
DBInstallationTest runs through several database upgrades and it has
been using the same database as other tests on a jenkins build. This bug
probably goes back to the initial implementation. It wasn't a problem
previously because most if not all tests run before DBInstallationTest.
To avoid causing interference with other tests, it should use its own
dedicated database. This commit includes a fix for postgresql but not
for oracle. Oracle will have to be addressed separately.