[BZ 771417] Refactor logic for enabling/disabling form fields
Form fields were not getting enabled because the top level CanvasItem in
which they were contained is initially set to be disabled. Even though
child components are correctly enabled/disabled as necessary, that
top-level CanvasItem was not getting enabled; consequently, its disabled
state was propagated down. It was pointed out to me that in the version
of SmartGWT that was in use when this class was implemented, it may have
been the case that there was a bug where the enabled/disabled property
change was not propagated down.