[BZ 771406] Fixing issue with table headers not rendering
The problem turned out to be that we were adding the ListGrid to its
parent twice via addMember, once in onInit and once in onDraw. Removing
the first call to addMember(listGrid) in onInit appears to have resolved
the issue. All tables where the headers were previously not rendering
are rendering now.