[BZ 772259] Do not reset the pane property when a tab is being hidden
I am not sure what has changed with respect to Tab and TabSet since
SmartGWT 2.4, but resetting the pane property on the Tab that is being
hidden results in the tab staying visible. This commit remedies the
problem by not resetting the pane when the tab is hidden, and instead
caching both the tab and its pane. Then when the tab is made visible
again, the pane is reset.