It seems we didn't really understand how testng was executing tests when we submitted many test classes in a run. For example, when we submit the server jar tests we assumed that the test classes would be run sequentially, in perhaps some random order, and that each would have its Before/AfterClass methods executed atthe start and end of each test class execution.
That is not true.

Although testng seems to run most tests from a test class together, it does
not seem to make that guarantee.  From latest traces it seems that the
test classes are, toa degree, pooled and:
1) you can not assume that all test methods for a test class will execute
   before execution of test methods from another test class.
2) you can not assume that multiple BeforeClass methods will not execute
   prior to an AfterClass. For example, this can happen:
     Class1:BeforeClass
     Class2:BeforeClass
     Class1:AfterClass
     Class2:AfterClass

   Put another way: You can only assume that BeforeClass and AfterClass
   execute in order, for a single test class. You can not predict the
   execution order of all Before|AfterClass methods for all test classes.

3) The same rules as described in 2 apply to Before|AfterGroups

What this means is that you must be very careful with what is done in
Before|AfterClass and BeforeAfter|Groups.  The work must not conflict
with work done in other test classes that will execute in the same test
suite.

This is very important when multiple test classes are sharing some sort
of test resource.

In general, it is much safer, although perhaps a bit slower, to use
@BeforeMethod() and @AfterMethod(alwaysRun=true).

This misunderstanding was particularly bad for us in the server jar tests,
where many test classes extend AbstractEJB3Test and utilize the jboss
mbean server under the covers, in the prepareXXX methods.  This caused
clashes described above.

So, for server jar testing this reworks things to better share the mbean
server, even among tests in the same class. And migrates all
AbstractEJB3Test.prepare/unprepare calls from Before|AfterClass|Groups to
Before|After|Method.

also - remove/suppress IDE warnings from touched test classes

Conflicts:

	modules/enterprise/server/itests/src/test/java/org/rhq/enterprise/server/drift/DriftServerTest.java