[BZ 786106] Wrap calls to obtain managers in privileged blocks so that 3rd callers can safely obtain them.
The StandardBindings put all the managers into the script context before
the script engine is initialized with the security measures which makes
the managers available inside the scripts. Java code that gets injected as
other params into the scripts (like the "ProxyFactory" (of class
ResourceClientFactory) would suffer from access control exceptions when
it tried to obtain some manager while being called from the script because
it would try to call the methods from the LocalClient to obtain the remote
interfaces directly, without a wrapping in a privileged block). Obtaining
the remote interfaces is a safe operation wrt the scripts and so can be
wrapped in privileged block so that any caller of the LocalClient can
have access to the regardless of the access control restrictions in place.