[Bug 773031 - Component start() called more than once without an intervening stop()]
A variety of auto-formatting changes took place, actual changes are in:
  InventoryManager.prepareResourceForActivation
  InventoryManager.activateResource

The resource component state could be STOPPED or STARTED.  There was a
large window while a component was actually starting that a call
to prepareResourceForActivation would happily allow the component to
again be activated. So, added an actual STARTING state that can block
redundant activation.

Additionally, if forcing reactivation ensure a STARTED component is
stopped before being restarted.

Also:
- remove unnecessary and bad lazy state check
- comment out some unused/debug code
- remove warnings

Conflicts:

	modules/core/plugin-container/src/main/java/org/rhq/core/pc/inventory/InventoryManager.java