[BZ 788638] Do not cache ManagedComponent objects to avoid high memory usage.
With this commit we no longer cache the ManagedComponent objects we get
back from the profile service. Even when we were caching them, we still
regularly fetched them for metric collections.

The caching was introduced to address high CPU utilization during
availability checks. The logic in
ManagedComponentComponent.getAvailability has been refactored so that we
do not necessarily fetch the component on every availability check.