Send agent change set ack
Updating MongoDBDriftServer.saveChangeSet to send the agent the change
set acknowledgement. Also reverting the version of morphia to 0.99 for
now. The 1.00 snapshot version now requires a newer version of the
MongoDB driver, and I am not ready just yet to upgrade. I need to do
some more testing to see what does and does not currently work.