Rounding out support for DriftChangeSetCriteria queries
This commit adds support for including/excluding drift entries from the
results and for sorting by version. There is no support yet for the
methods defined in BaseCriteria. I'll probably defer that to a later
time after the plugin can be considered functionally complete. I do not
think supporting those methods is necessary to be complete.