[BZ 781602] Adding logic to keep order of tabs consistent
The set of visible tabs can vary from one resource to another; however,
the order of tabs relative to one another is fixed. For example, the
Operations tab could appear as the 3rd or 4th tab, and it could be the
last tab. Note that it should always appear after the summary,
inventory, monitoring, and events tabs in that order.  So if the
monitoring and events tabs are not visible, then the order should be
summary, inventory, operations, etc.

This commit adds an internal data structure to TwoLevelTabSet, similar
to a singly linked list a tab is a list node. Each node maintains two
pointers - one to the next visible tab and one to the next actual tab.