Adding more tests for persisting a change set
Two new test methods have been added. One tests persisting an initial
change set where the content is already in the db. The other test
involves persisting a drift change set with a changed file and a removed
file, and the new version of the file is not in the db.

This commit also includes some refactoring test set up/helper methods.