[BZ 755703] Updating MeasurementManager to get live per minute metrics
This bug required changes both on the server and on the agent. For the
server side, the implementation of
MeasurementDataManagerBean.findLiveData had to be refactored to pass the
schedule ids to the agent. For the agent/plugin container,
MeausrementManager.getRealTimeMeasurementValue was refactored to check
the requests for TRENDSUP and TRENDSDOWN (i.e., per-minute) metrics. A
per-minute metric is calculated and included in the report instead of
the raw value as was previously done.