There was a problem with backfilling. Disabled resources should remain disabled despite an agent going down, or being suspected as down.
Fixing this issue brought to light inherent weaknesses in
the legacy "No Availability Records equals Unknown" special
case.  So, bit the bullet and we now ensure an initial
Availability and ResourceAvailability record exist at Resource
persist time. The initial Availability is UNKNOWN, with a start
time of 0 (epoch), indicating that we have no avail history for
the Resource.  The initial Availability will typically be
replaced fairly quickly, when an agent reports th first actual
avail for he resource.

But, the existance of the record makes several things easier to
handle and to understand, and hopefully makes sense to the
user as well.

Also, fixed both a legacy logic error and a legacy coding error
in mergeAvailabilityReport, when performing an "insert" of past
avail data.

Updated tests to handle the change in strategy.