[BZ 800026] Adding logic so that order of sub tabs remains consistent
This commit adds a data structure to SubTab that is very similar to the
data structure added to TwoLevelTab in commit 4ad367088. An interal list
of the actual, fixed order of tabs as well as the order or positions of
visible tabs. This list is used to determine at what position a sub tab
should be inserted when made visible again.