Adding method to purge orphaned drift file content
This commit adds the method purgeOrphanedContent to MongoDBDriftServer.
This method is not part of the drift server plugin API, but
purgeOrphanedDriftFiles is. I am not (yet) implementing the latter
because there are some race conditions that have to be dealt with, and I
am pushing that off for now. The plan is to expose purgeOrphanedContent
as a plugin operation to provide a manual way of purging content.