[BZ 802544] Fixing problem with null links
The reason the null links were showing up is because with the title
decoration, VLayouts were getting added to NavigationItems instead of
the view objects as previously done. The problem with this is that the
view classes implement BookmarkableView, but the VLayout getting added
to the NavigationItem does not. When the view is subsequently rendered,
the renderView method of the view was not getting invoked which
resulted in the URLs with null showing up in them.

This commit resolves the issue by adding the original view object to the
NavigationItem. Only when a NavigtionItem is selected to become the
current view is it then decorated while renderView still gets invoked.