[BZ 802544] Do not invalidate user session too soon
When streaming the response back to the client SetCallerInceptor cannot
invalidate the user's session right away because the result is a
callback object that may make additional secure SLSB calls. To remedy
this, I am wrapping the returned StreamingOutput with another instance
that invalidates the session after the output is written.