[BZ 802544] Refactoring drift compliance report to make details query more efficient
The drift compliance report is nearly identical to the inventory summary
report except that the former contains an additional column, "In
Compliance?". I have refactored InventorySummaryHandler a bit so that it
could be reused to the greatest extent possible for the drift compliance
report.

This commit also adds a resourceTypeIds filter to ResourceCriteria so
that the details query in DriftComplianceHandler can filter on resource
types. Since only a handful of types support drift, this is much more
efficient than fetching all resources in inventory and then filtering
out in memory those that do not support drift.