[BZ 802544] Big refactoring of PlatformSummaryPortlet
There was a lot of business logic embedded in UI code in
PlatformSummaryPorlet. Much of that same logic is also needed to
generate the platform utilization report. This commit moves all that
logic to the SLSB layer. With this changes comes a new SLSB,
PlatformUtilizationManagerBean. This SLSB encapsulates the logic to
fetch and aggregate the data needed for both the UI and the report.