[Bug 786416 - Map configuration properties not shown in correct order] See the BZ for more detail, the thinking here was somewhat involved. In short the ConfigPropDefs for Map props were not being assigned an order_index when parsed from the descriptor and persisted into the db. But code assumed it was set properly, using that field in comparators to determine ordering. The order in which properties are declared in the plugin descriptor should be honored, and now it is.
This commit includes a db-upgrade step that will apply ordering to existing
Map properties. This presents something of an issue for legacy databases that
may have Maps that have been updated over time given plugin changes. The
applied ordering may not be correct, but seeing that the current ordering
is also incorrect, we may have to deal with issues, if any, as they come
up.

Additionally, given that we have order_index we shouldn't be relying on
the @OrderBy annotation on the entity's map field. It's confusing, and
unreliable given the chance of meta-data-update. The order_index field
should be used when sorting properties is required and as such,
the annotation has been removed.

While adding tests for this it became clear that metadata update for
map properties was itself badly broken. This has also been fixed, and
now also sets order_index properly.