[BZ 802544] Initial attempt at adding checkbox column to inventory summary report
The suggestion was made during a UX meeting to add a checkbox column to
allow the user to select which types he wants to export. This is more of
a prototype at this point to see if we want to do this. The
functionality is a bit off. Before being able to toggle the state of the
checkbox, the row first needs to be selected. This in effect makes it
seem you have to click the checkbox twice initially. Once the row is
selected, it is not an issue.