[BZ 802544] Fixing couple errors with recent operations report
The fist problem was in the UI code. We were trying to add the table
action to the table too soon, resulting in an NPE. It needs to be done
in configureTable. The second problem was in RecentOperationsHandler. We
were passing the wrong string to OperationRequestStatus.valueOf. It
should be the enum constant itself.