[BZ 802544] Avoid lazy init exceptions
While testing the alert definitions report, I managed to generate some lazy
init exceptions. Calling AlertDefinitionCriteria.fetchResource(true) was
needed, but I still hit the exception with the call to resource.getParentResource().
Resource.parentResource is lazily loaded. I need to investigate some
more, but it seems that the transaction is committed and the persistence
context is closed before the results are written the response output
stream.

As an interim solution to work past this, I've created a local cache of
resources. When a resource is loaded into the cache its parent resource
is fetched as well. This is only temporary as it will not be performant
for larger reports.

I have also temporarily commented out the filtering in
RecentAlertHandler and RecentOperationsHandler for testing/demoing. The
filters are not yet being passed in the request URIs.