[BZ 802544] Adding initial support for selecting/deselecting all details
I spent a good bit of time investigating a good way to do this in the
column header, and unfortunately do not see one. You can easily add a
checkbox column that toggles on/off selection of all rows, but that is
not quite the behavior we want. Furthermore, I don't see an easy way to
add a label to that column. An easier option for the column header would
be to add another item to the export details column header. I'm not sure
how intuitive that would be. Adding the checkbox to the filters section
has the benefits of being easy to implement and being consistent with
where we place our other filter controls.