[BZ 802544]] Don't pass empty string for query params
The multi-valued query params in ExportModalWindow are stored as arrays.
The views build up the filter values in a list and call List.toArray.
When the list is empty though an array of length zero is passed to
ExportModalWindow. This was resulting in query params with the empty
string as values when we really want the defaults.