[BZ 802544] Refactorig inventory summary and drift compliance reports code
The inventory summary and drift compliance reports are nearly the same.
The drift compliance report limits the resource types to those that
support drift. It also adds an "in compliance" column. Since more
functionality needs to be added to these reports in support of exporting
details, I have decided to refactor the table and data source code into
common, shared classes so as to avoid having to duplicate the new logic.