[BZ 800453] Refactoring logic for including details
The requirements for the inventory summary and drift compliance reports
have changed a bit. This commit makes those changes. All of the resource
type rows are always included now. If one or more resource type ids are
specified, then the resource details for the types will be included in
the report as well. If the showAllDetails param is true, then all
resource details for all types will be included.