[BZ 800453] Fixing bug where map of install counts was not using unique keys
When details are included in an inventory summary report, the resource
install counts get stored in a map. The keys, which were just the
resource type ids, were not unique. You can have multiple
ResourceInstallCounts with the same resource type id. Now the key is a
combination of the type id and version.