[BZ 800453] Fixing date formatting so that we are only filtering on days
When selecting values from the date fields the time of the date was
defaulting to 12:00 pm. The reasoning for this is explained in this
smartgwt forum post
http://forums.smartclient.com/showthread.php?t=20550&highlight=DateItem.

We are only interesting on filtering on days, nothing more granular. So
the start date will always be set to 12:00 am of the date chosen, and
the end date will always be set to 11:59:59 of the date chosen.