[BZ 800453] Generate a url for the parent column
The alert definitions report in the UI displays a clickable field in the
parent column that sends you to the details page for the template
definition or the group definition for alert definitions created from
templates and groups respectively. The exported report has been modified
so that those URLs are now being written out to the parent column.