Related to Bug 786889 which deals with invalid (low) schedule intervals reaching the agent.  Although it seems unlikely that the bad interval was stored in the db, there exist ways to store a bad value and it can in fact reach the agent.  Don't let that happen by fixing the interval verification in the SLSB code and blocking all values < the minum collection interval. - Also, added test which failed to prove the issue, and now passes.
[Bug 786889 - Agent plug-in container should log warnings and disable metric collection for metrics with invalid collection intervals]
Catch any attempt to set a metric collection interval to an invalid value.
If found, set it to 20 minutes and report the problem in the log as an
ERROR, with a lot of supporting info to help track down the cause.

Move a private util method I had in ResourceContainer to ThrowableUtil
and use the new utility method.
(cherry-picked from a00eb5ee989cf5e0a91278ddf51f2ef1e7335b00)